requirements.txt
python-telegram-bot>=20.0
paypalrestsdk>=1.13.3
stripe>=5.5.0
psycopg2-binary>=2.9.5
redis>=4.5.0
python-dotenv>=0.19.2
sqlalchemy>=1.4.46
fastapi>=0.68.0
uvicorn>=0.15.0
cryptography>=3.4.8
reportlab>=3.6.0
requests>=2.28.0
email-validator>=1.3.0
pydantic>=1.10.0
jinja2>=3.1.0
pytest>=7.0.0
